Stage 2: flip pointer-cast sites from int32_t to intptr_t
Step 5 of the Stage 2 64-bit-clean sweep. Last step (long -> int32_t)
was over-eager: it renamed pointer-to-int casts like (long)vbuf or
(long)&v[i] to (int32_t), which is width-correct on 32-bit MSVC but
truncates on 64-bit. Flip those specific sites to intptr_t so the
types are honest everywhere.

Also widen the public voxsetframebuffer's first parameter from
int32_t to intptr_t in voxlap5.h and voxlap5.c, and simplify the
oracle's call from (int32_t)(intptr_t)g_fb to (intptr_t)g_fb.
Engine-internal `frameplace` stays int32_t for now — it feeds the
MSVC `_asm { add edx, frameplace }` blocks that Stage 3 will rewrite,
so there's no point re-typing it twice.

Hash-neutral on 32-bit MSVC: intptr_t and int32_t are both 4 bytes
there, so the emitted code is identical.

Preserved as-is (byte deref widening, not pointer casts):
  (int32_t)filptr[0], (int32_t)cptr[i], (int32_t)v[0] etc.
A sed-style audit picked out specifically addr-of operators and
known pointer identifiers (v, v2, vbuf, sptr, patbuf, xptr, xyptr,
zbuffermem, lptr).

Note: this does not widen int32_t variables that store pointer
values (frameplace, gpixy, col, pathashead, etc.). Those work on
32-bit because the values fit in 4 bytes; making them truly 64-bit
clean is Stage 3 territory when the inline asm goes away and
pointer-sized locals can be refactored.

+++ tests/oracle/oracle.c
@@ -153,7 +153,7 @@ int main(void) {
 	if (initvoxlap() < 0) { fprintf(stderr, "initvoxlap failed\n"); return 1; }
 
 	build_scene();
-	voxsetframebuffer((int32_t)(intptr_t)g_fb, BYTESPERLINE, XRES, YRES);
+	voxsetframebuffer((intptr_t)g_fb, BYTESPERLINE, XRES, YRES);
 
 	hf = fopen("hashes.txt", "w");
 	if (!hf) { fprintf(stderr, "cannot open hashes.txt\n"); return 2; }
